Separate out PerfectPrior.hpp
Also:
- Fix immediate assignment of variable values
