+++ wave_optimization/include/wave/optimization/factor_graph/Factor.hpp
@@ -6,14 +6,10 @@
 #ifndef WAVE_OPTIMIZATION_FACTOR_GRAPH_FACTOR_HPP
 #define WAVE_OPTIMIZATION_FACTOR_GRAPH_FACTOR_HPP
 
-#include <ceres/sized_cost_function.h>
-#include <memory>
-
 #include "wave/utils/math.hpp"
 #include "wave/optimization/factor_graph/FactorVariableBase.hpp"
 #include "wave/optimization/factor_graph/FactorBase.hpp"
 #include "wave/optimization/factor_graph/OutputMap.hpp"
-#include "wave/optimization/factor_graph/FactorMeasurement.hpp"
 
 
 namespace wave {
@@ -86,8 +82,6 @@ class Factor : public FactorBase {
                     MeasType measurement,
                     std::shared_ptr<VarTypes>... variable_ptrs);
 
-    ~Factor() override = default;
-
     int size() const override {
         return NumVars;
     }
@@ -123,66 +117,9 @@ class Factor : public FactorBase {
     VarArrayType variable_ptrs;
 };
 
-template <typename VarType>
-class PerfectPrior : public FactorBase {
-    using FactorType = PerfectPrior<VarType>;
-    using ViewType = typename VarType::ViewType;
-    using MeasType = FactorMeasurement<ViewType, void>;
-
- public:
-    constexpr static int NumVars = 1;
-    constexpr static int ResidualSize = MeasType::Size;
-    using ResidualType = Eigen::Matrix<double, ResidualSize, 1>;
-    using VarArrayType = FactorBase::VarVectorType;
-    using const_iterator = typename VarArrayType::const_iterator;
-
-    /** Construct with the given measurement and variable. */
-    explicit PerfectPrior(MeasType measurement,
-                          std::shared_ptr<VarType> variable_ptr)
-        : measurement{measurement}, variable_ptrs{variable_ptr} {
-        // Assign to the variable
-        variable_ptr->value.asVector() = measurement.value.asVector();
-    }
-
-    ~PerfectPrior() override = default;
-
-    int size() const override {
-        return NumVars;
-    }
-
-    int residualSize() const override {
-        return ResidualSize;
-    }
-
-    bool evaluateRaw(double const *const *, double *, double **) const
-      noexcept override {
-        return false;
-    }
-
-    /** Get a reference to the vector of variable pointers */
-    const VarVectorType &variables() const noexcept override {
-        return this->variable_ptrs;
-    }
-
-    /** Return true if this factor is a zero-noise prior */
-    bool isPerfectPrior() const noexcept override {
-        return true;
-    }
-
-    /** Print a representation for debugging. Used by operator<< */
-    void print(std::ostream &os) const override {}
-
- private:
-    /** Storage of the measurement */
-    MeasType measurement;
-
-    /** Pointers to the variables this factor is linked to */
-    VarArrayType variable_ptrs;
-};
-
 /** @} group optimization */
 }  // namespace wave
 
 #include "impl/Factor.hpp"
 
-#endif  // WAVE_OPTIMIZATION_FACTOR_GRAPH_PERFECT_PRIOR_HPP
+#endif  // WAVE_OPTIMIZATION_FACTOR_GRAPH_FACTOR_HPP
